+++ joblib/externals/loky/__init__.py
@@ -1,8 +1,9 @@
-r"""The :mod:`loky` module manages a pool of worker that can be reused across time.
+r"""The :mod:`loky` module manages a pool of worker that can be re-used across time.
 It provides a robust and dynamic implementation os the
 :class:`ProcessPoolExecutor` and a function :func:`get_reusable_executor` which
 hide the pool management under the hood.
 """
+
 from concurrent.futures import (
     ALL_COMPLETED,
     FIRST_COMPLETED,
@@ -41,4 +42,4 @@ __all__ = [
 ]
 
 
-__version__ = "3.4.1"
+__version__ = "3.5.2"